% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesSumarize.R
\name{genesSummarize}
\alias{genesSummarize}
\title{genesSumarize}
\usage{
genesSummarize(bamDataFrame, geneData = TAIR10_genes, chromosome = NULL,
  range = 0)
}
\description{
Function genesSumarize counts the coverage of each gene
}
